move peer service processor
