+++ ddtrace/tracer.py
@@ -17,7 +17,6 @@ from typing import TypeVar
 from typing import Union
 
 from ddtrace import config
-from ddtrace._trace.processor import PeerServiceProcessor
 from ddtrace._trace.processor import SpanAggregator
 from ddtrace._trace.processor import SpanProcessor
 from ddtrace._trace.processor import SpanSamplingProcessor
@@ -26,6 +25,7 @@ from ddtrace._trace.processor import TraceProcessor
 from ddtrace._trace.processor import TraceSamplingProcessor
 from ddtrace._trace.processor import TraceTagsProcessor
 from ddtrace.filters import TraceFilter
+from ddtrace.internal.peer_service.processor import PeerServiceProcessor
 from ddtrace.internal.processor.endpoint_call_counter import EndpointCallCounterProcessor
 from ddtrace.internal.sampling import SpanSamplingRule
 from ddtrace.internal.sampling import get_span_sampling_rules
